+++ CHANGELOG.md
@@ -5,8 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Changed
+
+- Rename `UnitName` module to `UnitLabel`
+- Rename `UnitName.getName()` to `UnitLabel.getLabel()`
+
 ## [1.0.0] - 2018-05-08
 
 ### Added
 
-* Initial Release
+- Initial Release
+
+[unreleased]: https://github.com/jonaskello/tslint-immutable/compare/v1.0.0...master
+[v1.0.0]: https://github.com/jonaskello/tslint-immutable/compare/v0.0.0...v1.0.0